reorganize package and remove NewAESWithNonceSize func
This reorganizes the package and changes some names
for future extensions. Further license headers are
added.
